requirements.txt
google-genai 
django 
